requirements.txt
requests==2.26.0
python-dotenv==0.19.2
Flask==2.2.5
gunicorn==20.1.0
Werkzeug==2.2.3
gtts==2.5.3

